% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_comparison.R
\name{calc_reconstruction_error_mixtureFA}
\alias{calc_reconstruction_error_mixtureFA}
\title{Compute SSE for a MixtureFA model}
\usage{
calc_reconstruction_error_mixtureFA(list_of_data, mfa_fit)
}
\arguments{
\item{list_of_data}{The data list.}

\item{mfa_fit}{The mixture FA result, which contains \code{$z}, \code{$Lambda[[k]]}, \code{$mu[[k]]}, \code{$Psi[[k]]}.}
}
\value{
SSE (sum of squared errors).
}
\description{
Uses the final hard assignments \code{mfa_fit$z}, reconstruct each subject's data.
}
