% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_comparison.R
\name{calc_reconstruction_error_mixturePCA}
\alias{calc_reconstruction_error_mixturePCA}
\title{Compute SSE for a MixturePCA model}
\usage{
calc_reconstruction_error_mixturePCA(list_of_data, mpca_fit)
}
\arguments{
\item{list_of_data}{The data list.}

\item{mpca_fit}{The mixture PCA result, containing \code{$z} and each cluster's \code{P, mu}.}
}
\value{
SSE (sum of squared errors).
}
\description{
Compute SSE for a MixturePCA model
}
