% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_comparison.R
\name{calc_reconstruction_error_singleFA}
\alias{calc_reconstruction_error_singleFA}
\title{Compute Total Reconstruction SSE for Single FA}
\usage{
calc_reconstruction_error_singleFA(list_of_data, fa_res)
}
\arguments{
\item{list_of_data}{The original data list.}

\item{fa_res}{A result from \code{fit_single_factor_analysis}.}
}
\value{
A numeric value of the sum of squared errors (SSE).
}
\description{
Compute Total Reconstruction SSE for Single FA
}
