% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_comparison.R
\name{calc_reconstruction_error_singlePCA}
\alias{calc_reconstruction_error_singlePCA}
\title{Compute Total Reconstruction SSE for Single PCA}
\usage{
calc_reconstruction_error_singlePCA(list_of_data, pca_res)
}
\arguments{
\item{list_of_data}{The data list.}

\item{pca_res}{The result from \code{fit_single_pca}.}
}
\value{
SSE (sum of squared errors).
}
\description{
Compute Total Reconstruction SSE for Single PCA
}
