% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bic.R
\name{compute_bic}
\alias{compute_bic}
\title{Compute BIC for MFA or Mixture PCA}
\usage{
compute_bic(logLik, K, r, M, N_total_rows)
}
\arguments{
\item{logLik}{The log-likelihood value.}

\item{K}{Number of clusters.}

\item{r}{Factor or principal component dimension.}

\item{M}{Number of observed channels.}

\item{N_total_rows}{Total number of rows across all data (used for \code{log(N)}).}
}
\value{
A numeric value for BIC.
}
\description{
This function computes a simple BIC measure given a log-likelihood and
a rough count of model parameters, assuming diagonal noise.
}
\details{
The number of parameters is counted as:
\deqn{K*(M*r + 2M) + (K-1).}
}
\examples{
bic_val <- compute_bic(-1234.5, K=3, r=4, M=8, N_total_rows=500)
}
