% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_BIC_mpca}
\alias{compute_BIC_mpca}
\title{Compute BIC for a Mixture PCA Model}
\usage{
compute_BIC_mpca(loglik, K, r, M, N)
}
\arguments{
\item{loglik}{The total mixture log-likelihood from \code{compute_logLik_mpca}.}

\item{K}{Number of clusters.}

\item{r}{Dimension of principal components.}

\item{M}{Observed dimension.}

\item{N}{Number of subjects.}
}
\value{
A numeric BIC value.
}
\description{
Given the log-likelihood from \code{\link{compute_logLik_mpca}}, and the
number of clusters \code{K}, dimension \code{r}, etc., this function
computes a naive BIC:
\deqn{ \mathrm{BIC} = -2 \log \mathcal{L} + \nu \log(N), }
where \(\nu\) is a parameter count for MPCA. One might use
\code{K*(M*r + M + 1) + (K-1)} as a rough approximation, i.e. for each cluster
we have \code{(M*r)} for the directions, \code{M} for the mean, and \code{1}
for \(\sigma^2\). Then \code{(K-1)} for mixing proportions.
}
