% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_cluster_sizes}
\alias{compute_cluster_sizes}
\title{Compute Cluster Sizes from a Fitted MFA (or Mixture PCA) Model}
\usage{
compute_cluster_sizes(res)
}
\arguments{
\item{res}{A fitted mixture model object that contains \code{z}.}
}
\value{
A table of cluster frequencies.
}
\description{
Given a result with a vector \code{z} of length N (the cluster assignments),
this function returns a frequency table, i.e., how many subjects are assigned
to each cluster.
}
\examples{
\dontrun{
tab <- compute_cluster_sizes(fit_mfa)
print(tab)
}
}
