% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_cluster_sizes_mpca}
\alias{compute_cluster_sizes_mpca}
\title{Compute Cluster Sizes for Mixture PCA}
\usage{
compute_cluster_sizes_mpca(res)
}
\arguments{
\item{res}{A fitted mixture PCA model result, containing \code{$z}.}
}
\value{
A table of frequencies of cluster assignments.
}
\description{
Like \code{\link{compute_cluster_sizes}}, but for a Mixture PCA model result.
}
