% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_factor_scores}
\alias{compute_factor_scores}
\title{Compute Factor Scores for a Single MFA Cluster}
\usage{
compute_factor_scores(X, mu, Lambda, Psi)
}
\arguments{
\item{X}{A numeric matrix of size \code{(T x M)}.}

\item{mu}{A length-M mean vector.}

\item{Lambda}{An \code{M x r} loading matrix.}

\item{Psi}{An \code{M x M} diagonal matrix for unique variances.}
}
\value{
A \code{(T x r)} matrix of factor scores.
}
\description{
Given a single cluster's parameters (\code{mu, Lambda, Psi}) and data \code{X},
this function returns the estimated factor scores.
}
\details{
This follows the standard posterior mean formula in factor analysis:
\eqn{\mathrm{E}[Z|X] = (I + \Lambda^T \Psi^{-1} \Lambda)^{-1} \Lambda^T \Psi^{-1} (X - \mu)}.
}
\examples{
\dontrun{
# Suppose we have X (T x M), mu (M), Lambda (M x r), Psi (M x M).
scores <- compute_factor_scores(X, mu, Lambda, Psi)
head(scores)
}

}
\seealso{
\code{\link{compute_global_vaf_mfa}}
}
