% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_comparison.R
\name{fit_single_factor_analysis}
\alias{fit_single_factor_analysis}
\title{Fit a Single Factor Analysis (Ignoring Clusters)}
\usage{
fit_single_factor_analysis(list_of_data, r = 2)
}
\arguments{
\item{list_of_data}{A list of (T_i x M) matrices.}

\item{r}{The factor dimension.}
}
\value{
A list with elements \code{(loadings, mu, diagPsi, logLik, BIC)}.
}
\description{
This function concatenates all subjects' data into one large matrix,
then performs a naive factor analysis. It returns a simple BIC measure
and the fitted loadings, mean, and diagonal Psi.
}
