% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_model_by_K_r}
\alias{get_model_by_K_r}
\title{Retrieve a Specific Model by (K, r) from a Grid Search}
\usage{
get_model_by_K_r(selection_obj, K_target, r_target)
}
\arguments{
\item{selection_obj}{The object returned by \code{\link{select_optimal_K_r_mfa}}.}

\item{K_target}{Desired number of clusters.}

\item{r_target}{Desired factor dimension.}
}
\value{
The fitted model (the same structure returned by \code{mfa_em_fit}),
  or \code{NULL} if no match found.
}
\description{
After calling \code{\link{select_optimal_K_r_mfa}} (or a similar function),
you have \code{$all_models} containing fits for various \code{(K, r)}. This
function extracts the model for a user-specified pair \code{(K_target, r_target)}.
}
\examples{
\dontrun{
sel_obj <- select_optimal_K_r_mfa(my_data, Kvec=2:4, rvec=1:3)
best_model <- get_model_by_K_r(sel_obj, K_target=3, r_target=2)
}

}
