% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_model_by_K_r_mpca}
\alias{get_model_by_K_r_mpca}
\title{Retrieve a Mixture PCA Model by (K, r)}
\usage{
get_model_by_K_r_mpca(selection_obj, K_target, r_target)
}
\arguments{
\item{selection_obj}{The output of \code{\link{select_optimal_K_r_mpca}}.}

\item{K_target}{Desired K.}

\item{r_target}{Desired r.}
}
\value{
The fitted model (same structure as from \code{mixture_pca_em_fit}), or \code{NULL}.
}
\description{
Given a \code{selection_obj} from \code{\link{select_optimal_K_r_mpca}},
extracts the model for the specified \code{(K_target, r_target)} if it exists.
}
