% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tables.R
\name{make_table1}
\alias{make_table1}
\title{Create a Static Table of Simulation Conditions (Table 1)}
\usage{
make_table1()
}
\value{
A \code{data.frame} with columns \code{Parameter} and \code{Values}.
}
\description{
This function returns a small \code{data.frame} describing the parameter ranges
used in the simulation (K, r, N, M, etc.). Typically printed via \code{kable()} in a vignette or paper.
}
\examples{
\dontrun{
df_table1 <- make_table1()
knitr::kable(df_table1, caption="Table 1: Conditions")
}

}
