% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_all_reconstructed_waveforms_mpca}
\alias{plot_all_reconstructed_waveforms_mpca}
\title{Plot Reconstructed Waveforms for All Clusters and Channels (MPCA)}
\usage{
plot_all_reconstructed_waveforms_mpca(
  mpca_fit,
  list_of_data,
  overlay_subjects = TRUE
)
}
\arguments{
\item{mpca_fit}{A fitted MPCA model (with z, W, mu, etc.).
Also \code{factor_scores[[i]]} must be available if we want to reconstruct Xhat.}

\item{list_of_data}{A list of (T_i x M) data matrices used in fitting.}

\item{overlay_subjects}{Logical; if TRUE, each subject in the same cluster is drawn
in the same facet. If FALSE, the function computes the average waveforms across
subjects in each cluster.}
}
\value{
A ggplot object with line plots of reconstructed waveforms,
  facet by cluster & channel.
}
\description{
This function plots the reconstructed waveforms (Xhat_i) for each subject
across all channels, grouped by cluster. By default, it overlays all subjects
in the same cluster, then uses facet to separate by (Cluster, Channel).
}
\examples{
\dontrun{
  mpca_fit <- compute_factor_scores_mpca(mpca_fit, list_of_data)
  p_recon <- plot_all_reconstructed_waveforms_mpca(mpca_fit, list_of_data)
  print(p_recon)
}
}
