% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{plot_fig1_ARI}
\alias{plot_fig1_ARI}
\title{Plot ARI vs. Separation/Noise (Figure 1)}
\usage{
plot_fig1_ARI(df)
}
\arguments{
\item{df}{A \code{data.frame} with columns \code{K_true, r_true, N, M, sep, noise, ARI_MFA, ARI_PCA}.}
}
\value{
A \code{ggplot} object.
}
\description{
Given a result data frame with columns like \code{ARI_MFA, ARI_PCA, sep, noise},
this function reshapes them into long format (MFA vs. MPCA) and plots ARI vs. separation,
color-coded by noise, using \code{ggplot2}.
}
\examples{
\dontrun{
p <- plot_fig1_ARI(df_merged)
ggsave("figure1_ARI_sep_noise.pdf", p)
}

}
