% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{plot_fig2_BIC_heatmap}
\alias{plot_fig2_BIC_heatmap}
\title{Plot BIC Heatmap for a Subset (Figure 2)}
\usage{
plot_fig2_BIC_heatmap(df)
}
\arguments{
\item{df}{A \code{data.frame} with columns including \code{BIC_MFA, K_true, r_true}.}
}
\value{
A \code{ggplot} object.
}
\description{
This function filters the data for a specific condition, e.g. \code{N=50, M=8, sep=1.0, noise=1.0},
and plots a heatmap of \code{BIC_MFA} with \code{r_true} on the x-axis and \code{K_true} on the y-axis.
}
\examples{
\dontrun{
p <- plot_fig2_BIC_heatmap(df_merged)
ggsave("figure2_BIC_heatmap.pdf", p)
}

}
