% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{plot_fig4_loadings_heatmap}
\alias{plot_fig4_loadings_heatmap}
\title{Plot Factor Loadings Heatmap (Figure 4)}
\usage{
plot_fig4_loadings_heatmap(mfa_fit, cluster_id = 1)
}
\arguments{
\item{mfa_fit}{A fitted MFA model with \code{$Lambda}.}

\item{cluster_id}{Which cluster's loadings to plot (1..K).}
}
\value{
A \code{ggplot} object.
}
\description{
Given an MFA fit object containing \code{Lambda}, this function plots a heatmap
for the loadings in a specific cluster.
}
\examples{
\dontrun{
fit_mfa <- readRDS("some_fit.rds")
p <- plot_fig4_loadings_heatmap(fit_mfa, cluster_id=1)
ggsave("figure4_loadings_heatmap.pdf", p)
}

}
