% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_waveform_comparison_mfa}
\alias{plot_waveform_comparison_mfa}
\title{Plot Original vs. Reconstructed Waveform for an MFA Model}
\usage{
plot_waveform_comparison_mfa(mfa_fit, list_of_data, subject_id, channel = 1)
}
\arguments{
\item{mfa_fit}{A fitted MFA model with \code{Lambda, Psi, mu, z}.}

\item{list_of_data}{The original data list used to fit the model.}

\item{subject_id}{Index of the subject to plot.}

\item{channel}{Which channel (muscle) to plot. Defaults to 1.}
}
\value{
A \code{ggplot} object showing original and reconstructed waveforms.
}
\description{
Plot Original vs. Reconstructed Waveform for an MFA Model
}
