% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{posthoc_mfa_evaluation}
\alias{posthoc_mfa_evaluation}
\title{Post-hoc Evaluation of MFA Grid Search}
\usage{
posthoc_mfa_evaluation(list_of_data, selection_obj)
}
\arguments{
\item{list_of_data}{The same list of data used to fit the models.}

\item{selection_obj}{The object returned by \code{\link{select_optimal_K_r_mfa}}.}
}
\value{
A data frame with columns:
\item{K}{Number of clusters.}
\item{r}{Factor dimension.}
\item{logLik}{Log-likelihood from the fit.}
\item{BIC}{BIC from the fit.}
\item{GlobalVAF}{Global variance accounted for, \eqn{1 - \mathrm{SSE}/\mathrm{SST}}.}
\item{minClusterSize}{Minimum cluster size.}
\item{maxClusterSize}{Maximum cluster size.}
}
\description{
Given the output of \code{\link{select_optimal_K_r_mfa}}, this function computes
additional metrics for each fitted model in \code{all_models}, such as
global VAF (\code{\link{compute_global_vaf_mfa}}) and cluster sizes
(\code{\link{compute_cluster_sizes}}).
}
\details{
This function loops over each entry in \code{selection_obj$all_models},
extracts the model, calls \code{\link{compute_global_vaf_mfa}} and
\code{\link{compute_cluster_sizes}}, and records the results in a combined table.
}
\examples{
\dontrun{
sel_obj <- select_optimal_K_r_mfa(my_data, Kvec=2:4, rvec=1:3)
df_posthoc <- posthoc_mfa_evaluation(my_data, sel_obj)
head(df_posthoc)
}

}
\seealso{
\code{\link{select_optimal_K_r_mfa}}, \code{\link{compute_factor_scores}}
}
