% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{posthoc_mpca_evaluation}
\alias{posthoc_mpca_evaluation}
\title{Post-hoc Evaluation of Mixture PCA Grid Search}
\usage{
posthoc_mpca_evaluation(list_of_data, selection_obj)
}
\arguments{
\item{list_of_data}{The same list of data used for the grid search.}

\item{selection_obj}{The object returned by \code{\link{select_optimal_K_r_mpca}}.}
}
\value{
A data frame with columns \code{K, r, logLik, BIC, GlobalVAF, minClusterSize, maxClusterSize}.
}
\description{
Takes the output of \code{\link{select_optimal_K_r_mpca}} and computes additional
metrics (global VAF, cluster sizes, etc.) for each entry in \code{all_models}.
}
