% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_visualize_all.R
\name{read_and_visualize_all}
\alias{read_and_visualize_all}
\title{Read All Results and Produce Tables/Figures}
\usage{
read_and_visualize_all(result_dir = "results")
}
\arguments{
\item{result_dir}{Directory containing .rds result files (default "results").}
}
\value{
Invisibly returns the merged data frame of all results.
}
\description{
This function demonstrates how to read all .rds files from a directory,
merge them into a single \code{data.frame}, generate a couple of tables (Table 1, Table 2),
and produce various figures (Figure 1, Figure 2, etc.) saving them to files.
}
\details{
The user can modify the conditions in \code{\link{make_table1}} and \code{\link{make_table2}}
to match their actual experiment ranges. Figures are generated via \code{ggsave}.
}
\examples{
\dontrun{
df_all <- read_and_visualize_all("results")
}

}
