% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{reconstruct_singleMPCA}
\alias{reconstruct_singleMPCA}
\title{Reconstruct a single subject's data from a Mixture PCA model}
\usage{
reconstruct_singleMPCA(X, mu, W, Z)
}
\arguments{
\item{X}{Original \code{(T x M)} data}

\item{mu}{An \code{(M)} mean vector}

\item{W}{An \code{(M x r)} loading matrix}

\item{Z}{A \code{(T x r)} factor score matrix}
}
\value{
A \code{(T x M)} reconstructed matrix Xhat
}
\description{
Reconstruct a single subject's data from a Mixture PCA model
}
