% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_parameters.R
\name{run_main_experiment_parallel_save}
\alias{run_main_experiment_parallel_save}
\title{Main Example: Run Parallel MFA/MPCA Simulation and Save Each Condition}
\usage{
run_main_experiment_parallel_save()
}
\value{
The merged data frame of results across all conditions.
}
\description{
Main Example: Run Parallel MFA/MPCA Simulation and Save Each Condition
}
\examples{
\dontrun{
df_out <- run_main_experiment_parallel_save()
head(df_out)
}
}
