% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_parameters.R
\name{compute_bic}
\alias{compute_bic}
\title{Compute BIC for MFA or Mixture PCA}
\usage{
compute_bic(logLik, K, r, M, N_total_rows)
}
\arguments{
\item{logLik}{The log-likelihood value.}

\item{K}{Number of clusters.}

\item{r}{Factor or principal component dimension.}

\item{M}{Number of observed channels.}

\item{N_total_rows}{Total number of rows across all data (used for log(N)).}
}
\description{
Compute BIC for MFA or Mixture PCA
}
