% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_parameters.R
\name{plot_fig1_ARI}
\alias{plot_fig1_ARI}
\title{Plot ARI vs. Separation/Noise (Figure 1)}
\usage{
plot_fig1_ARI(df)
}
\arguments{
\item{df}{A \code{data.frame}.}
}
\value{
A ggplot object.
}
\description{
Plot ARI vs. Separation/Noise (Figure 1)
}
