% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_parameters.R
\name{plot_fig2_BIC_heatmap}
\alias{plot_fig2_BIC_heatmap}
\title{Plot BIC Heatmap for a Subset (Figure 2)}
\usage{
plot_fig2_BIC_heatmap(df)
}
\arguments{
\item{df}{A \code{data.frame}.}
}
\value{
A ggplot object.
}
\description{
Plot BIC Heatmap for a Subset (Figure 2)
}
