% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{plot_fig3_init_boxplot}
\alias{plot_fig3_init_boxplot}
\title{Plot Boxplot of ARI for Multiple Initial Seeds (Figure 3)}
\usage{
plot_fig3_init_boxplot(df_multiple)
}
\arguments{
\item{df_multiple}{A \code{data.frame} with columns \code{ARI_MFA, ARI_PCA}.}
}
\value{
A \code{ggplot} object.
}
\description{
If the data frame contains multiple runs with the same parameters but different
initial seeds, this function can show a boxplot of ARI distribution for MFA and MPCA.
}
