% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_parameters.R
\name{plot_fig4_loadings_heatmap}
\alias{plot_fig4_loadings_heatmap}
\title{Plot Factor Loadings Heatmap (Figure 4)}
\usage{
plot_fig4_loadings_heatmap(mfa_fit, cluster_id = 1)
}
\arguments{
\item{mfa_fit}{A fitted MFA model with \code{$Lambda}.}

\item{cluster_id}{Which cluster's loadings to plot (1..K).}
}
\description{
Plot Factor Loadings Heatmap (Figure 4)
}
