% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_main_experiment_parallel_save.R
\name{run_main_experiment_parallel_save}
\alias{run_main_experiment_parallel_save}
\title{Main Example: Run Parallel MFA/MPCA Simulation and Save Each Condition}
\usage{
run_main_experiment_parallel_save()
}
\value{
The merged data frame of results across all conditions.
}
\description{
This function calls \code{\link{simulate_and_run_parallel_save}} with some default arguments,
prints the merged results, and generates a simple figure comparing ARI_MFA and ARI_PCA.
}
\examples{
\dontrun{
df_out <- run_main_experiment_parallel_save()
head(df_out)
}
}
